passage du Float au BigDecimal et exceptionHandler
nettoyage imports

+++ src/main/java/com/currencyconverter/api/entity/Rates.java
@@ -1,5 +1,6 @@
 package com.currencyconverter.api.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
@@ -7,9 +8,9 @@ public class Rates {
 	
 	private final String base;
 	private final LocalDate date;
-	private final Map<String, Float> rates;
+	private final Map<String, BigDecimal> rates;
 	
-	public Rates(String base, LocalDate date, Map<String, Float> rates) {
+	public Rates(String base, LocalDate date, Map<String, BigDecimal> rates) {
 		super();
 		this.base = base;
 		this.date = date;
@@ -22,7 +23,7 @@ public class Rates {
 	public LocalDate getDate() {
 		return date;
 	}
-	public Map<String, Float> getRates() {
+	public Map<String, BigDecimal> getRates() {
 		return rates;
 	}
